refactor: DataStore에 관한 ë¶분 부분 클래스로 책임 분리
