+++ android/app/src/main/java/com/teameetmeet/meetmeet/data/datasource/LocalCalendarDataSource.kt
@@ -1,7 +1,7 @@
 package com.teameetmeet.meetmeet.data.datasource
 
-import com.teameetmeet.meetmeet.data.database.dao.EventDao
-import com.teameetmeet.meetmeet.data.database.entity.Event
+import com.teameetmeet.meetmeet.data.local.database.dao.EventDao
+import com.teameetmeet.meetmeet.data.local.database.entity.Event
 import com.teameetmeet.meetmeet.data.toDateLong
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
