+++ android/app/src/main/java/com/teameetmeet/meetmeet/data/local/database/AppDatabase.kt
@@ -1,9 +1,9 @@
-package com.teameetmeet.meetmeet.data.database
+package com.teameetmeet.meetmeet.data.local.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.teameetmeet.meetmeet.data.database.dao.EventDao
-import com.teameetmeet.meetmeet.data.database.entity.Event
+import com.teameetmeet.meetmeet.data.local.database.dao.EventDao
+import com.teameetmeet.meetmeet.data.local.database.entity.Event
 
 @Database(entities = [Event::class], version = 0)
 abstract class AppDatabase : RoomDatabase() {
