+++ android/app/src/main/java/com/teameetmeet/meetmeet/data/local/datastore/DataStoreHelper.kt
@@ -1,0 +1,32 @@
+package com.teameetmeet.meetmeet.data.local.datastore
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.edit
+import androidx.datastore.preferences.core.stringPreferencesKey
+import com.teameetmeet.meetmeet.data.repository.UserRepository
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import javax.inject.Inject
+
+class DataStoreHelper @Inject constructor(
+    private val dataStore: DataStore<Preferences>
+) {
+
+    suspend fun storeAppToken(accessToken: String, refreshToken: String) {
+        dataStore.edit{
+            it[ACCESS_TOKEN] = accessToken
+            it[REFRESH_TOKEN] = refreshToken
+        }
+    }
+
+    fun getAppToken(): Flow<String?> =
+        dataStore.data.map { it[ACCESS_TOKEN] }
+
+
+
+    companion object {
+        val ACCESS_TOKEN = stringPreferencesKey("accessToken")
+        val REFRESH_TOKEN = stringPreferencesKey("refreshToken")
+    }
+}
