+++ android/app/src/main/java/com/teameetmeet/meetmeet/data/local/datastore/di/DataStoreModule.kt
@@ -1,4 +1,4 @@
-package com.teameetmeet.meetmeet.data.di
+package com.teameetmeet.meetmeet.data.local.datastore.di
 
 import android.content.Context
 import androidx.datastore.core.DataStore
@@ -11,6 +11,7 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
+import com.teameetmeet.meetmeet.data.local.datastore.DataStoreHelper
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -22,6 +23,10 @@ class DataStoreModule {
         PreferenceDataStoreFactory.create(
             produceFile = { context.preferencesDataStoreFile(PREFERENCES_STORE_NAME) }
         )
+
+    @Provides
+    @Singleton
+    fun provideDataStore(dataStore: DataStore<Preferences>): DataStoreHelper = DataStoreHelper(dataStore)
 
     companion object {
         private const val PREFERENCES_STORE_NAME = "AppDataStore"
